Split out `isExceedingSiteMaxUploadSize`
