+++ client/lib/media/utils/index.js
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import path from 'path';
-import { includes, omitBy, startsWith, get } from 'lodash';
+import { includes, omitBy } from 'lodash';
 import debugFactory from 'debug';
 const debug = debugFactory( 'calypso:media' );
 
@@ -19,7 +19,6 @@ import {
 } from 'lib/media/constants';
 import { stringify } from 'lib/shortcode';
 import impureLodash from 'lib/impure-lodash';
-import versionCompare from 'lib/version-compare';
 import { getUrlParts } from 'lib/url';
 import wpcom from 'lib/wp';
 import { getEditorPostId } from 'state/ui/editor/selectors';
@@ -28,6 +27,7 @@ import { getFileExtension } from 'lib/media/utils/get-file-extension';
 import { getMimeType } from 'lib/media/utils/get-mime-type';
 import { isSupportedFileTypeInPremium } from 'lib/media/utils/is-supported-file-type-in-premium';
 import { isSupportedFileTypeForSite } from 'lib/media/utils/is-supported-file-type-for-site';
+import { isExceedingSiteMaxUploadSize } from 'lib/media/utils/is-exceeding-site-max-upload-size';
 
 const { uniqueId } = impureLodash;
 
@@ -46,39 +46,7 @@ export { isSiteAllowedFileTypesToBeTrusted } from 'lib/media/utils/is-site-allow
 export { getAllowedFileTypesForSite } from 'lib/media/utils/get-allowed-file-types-for-site';
 export { isSupportedFileTypeInPremium } from 'lib/media/utils/is-supported-file-type-in-premium';
 export { isSupportedFileTypeForSite } from 'lib/media/utils/is-supported-file-type-for-site';
-
-/**
- * Returns true if the specified item exceeds the maximum upload size for
- * the given site. Returns null if the bytes are invalid, the max upload
- * size for the site is unknown or a video is being uploaded for a Jetpack
- * site with VideoPress enabled. Otherwise, returns true.
- *
- * @param  {object}   item  Media object
- * @param  {object}   site  Site object
- * @returns {?boolean}       Whether the size exceeds the site maximum
- */
-export function isExceedingSiteMaxUploadSize( item, site ) {
-	const bytes = item.size;
-
-	if ( ! site || ! site.options ) {
-		return null;
-	}
-
-	if ( ! isFinite( bytes ) || ! site.options.max_upload_size ) {
-		return null;
-	}
-
-	if (
-		site.jetpack &&
-		includes( get( site, 'options.active_modules' ), 'videopress' ) &&
-		versionCompare( get( site, 'options.jetpack_version' ), '4.5', '>=' ) &&
-		startsWith( getMimeType( item ), 'video/' )
-	) {
-		return null;
-	}
-
-	return bytes > site.options.max_upload_size;
-}
+export { isExceedingSiteMaxUploadSize } from 'lib/media/utils/is-exceeding-site-max-upload-size';
 
 /**
  * Returns true if the provided media object is a VideoPress video item.
